Add files via upload
Sorry, I noticed that the movement correction broke if the search window moved off the boundary of the image.... 
I had to make a change to `findCOMRefObj.m` to avoid problems when refPoints search region drifts off the edge of the image for  i=2:size(xyz,1) refPoints. It's the same fix that you already had for the i=1 window.